requirements.txt
Flask==3.1.0
openpyxl==3.1.5
pdfminer.six==20231228
sentence-transformers==3.4.1
scikit-learn==1.6.1
requests==2.32.3
beautifulsoup4==4.12.3
